+++ python/lsst/ts/salkafka/component_producer_set.py
@@ -25,6 +25,7 @@ __all__ = ["ComponentProducerSet"]
 import argparse
 import asyncio
 import concurrent.futures
+import copy
 import functools
 import logging
 import multiprocessing
@@ -37,6 +38,7 @@ import sys
 import traceback
 
 from lsst.ts import salobj
+from lsst.ts import utils
 from .topic_names_set import TopicNamesSet
 from .component_producer import ComponentProducer
 from .kafka_producer_factory import KafkaConfiguration, KafkaProducerFactory
@@ -162,9 +164,9 @@ class ComponentProducerSet:
         # This is different than start_task in order to avoid a race condition
         # when code (typically a unit test) creates a ComponentProducerSet
         # and immediately starts waiting for start_task to be done.
-        self._interruptable_start_task = salobj.make_done_future()
+        self._interruptable_start_task = utils.make_done_future()
 
-        self._run_producer_subprocess_task = salobj.make_done_future()
+        self._run_producer_subprocess_task = utils.make_done_future()
 
         # A task that ends when the service is interrupted.
         self._wait_forever_task = asyncio.Future()
@@ -257,7 +259,8 @@ class ComponentProducerSet:
         Parameters
         ----------
         kafka_config : `KafkaConfig`
-            Kafka configuration.
+            Kafka configuration. The ``partitions`` field is ignored,
+            in favor of the value in ``topic_names``.
         component : `str`
             Name of a SAL component for which to handle a subset of topics.
         index : `int`
@@ -274,6 +277,8 @@ class ComponentProducerSet:
             Length of the DDS read queue. Must be greater than or equal to
             `salobj.domain.DDS_READ_QUEUE_LEN`, which is the default.
         """
+        kafka_config = copy.copy(kafka_config)
+        kafka_config.partitions = topic_names.partitions
         producer_set = cls(kafka_config=kafka_config, log_level=log_level)
 
         producer_set._run_producer_subprocess_task = asyncio.create_task(
@@ -321,7 +326,9 @@ class ComponentProducerSet:
             "--partitions",
             type=int,
             default=1,
-            help="Number of partitions for each Kafka topic.",
+            help="Number of partitions for each Kafka topic. "
+            "A file specified as the --file argument may override this value "
+            "for each set of topics.",
         )
         parser.add_argument(
             "--replication-factor",
@@ -570,7 +577,18 @@ class ComponentProducerSet:
         queue_len : `int`, optional
             Length of the DDS read queue. Must be greater than or equal to
             `salobj.domain.DDS_READ_QUEUE_LEN`, which is the default.
+
+        Raises
+        ------
+        ValueError
+            If topic_names.partitions != self.kafka_config.partitions.
         """
+        if topic_names.partitions != self.kafka_config.partitions:
+            raise ValueError(
+                f"topic_names.partitions={topic_names.partitions} != "
+                f"kafka_config.partitions={self.kafka_config.partitions}"
+            )
+
         loop = asyncio.get_running_loop()
         for s in (signal.SIGTERM, signal.SIGINT, signal.SIGHUP):
             loop.add_signal_handler(s, self.signal_handler)