+++ python/lsst/ts/salkafka/topic_names_set.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 # This file is part of ts_salkafka.
 #
 # Developed for the LSST Telescope and Site Systems.
@@ -21,6 +23,8 @@
 
 __all__ = ["TopicNames", "TopicNamesSet"]
 
+import dataclasses
+import collections.abc
 import yaml
 
 import jsonschema
@@ -28,7 +32,10 @@ import jsonschema
 from lsst.ts import idl
 from lsst.ts import salobj
 
+TOPIC_CATEGORIES = ("commands", "events", "telemetry")
 
+
+@dataclasses.dataclass
 class TopicNames:
     """A collection of topic names.
 
@@ -38,27 +45,31 @@ class TopicNames:
 
     Parameters
     ----------
+    partitions : `int`
+        The desired number of Kafka partitions for these topics.
     add_ackcmd : `bool`, optional
         Add ``ackcmd`` topic to the producer?
     commands : `list` of `str`, optional
         Commands to add to the producer, with no prefix, e.g. "enable".
+        Converted to a sorted list.
     events : `list` of `str`, optional
         Events to add to the producer, with no prefix, e.g. "summaryState".
+        Converted to a sorted list.
     telemetry : `list` of `str`, optional
         Telemtry topics to add to the producer.
+        Converted to a sorted list.
     """
 
-    def __init__(
-        self,
-        add_ackcmd=False,
-        commands=(),
-        events=(),
-        telemetry=(),
-    ):
-        self.add_ackcmd = bool(add_ackcmd)
-        self.commands = sorted(commands)
-        self.events = sorted(events)
-        self.telemetry = sorted(telemetry)
+    partitions: int
+    add_ackcmd: bool = False
+    commands: collections.abc.Sequence[str] = ()
+    events: collections.abc.Sequence[str] = ()
+    telemetry: collections.abc.Sequence[str] = ()
+
+    def __post_init__(self):
+        for category in TOPIC_CATEGORIES:
+            sorted_list = sorted(getattr(self, category))
+            setattr(self, category, sorted_list)
 
 
 class TopicNamesSet:
@@ -82,6 +93,8 @@ class TopicNamesSet:
         List of `TopicNames` entries. Need not be complete, but if any topics
         are missing, the ``topic_names_list`` attribute has one additional item
         that specifies all remaining topics.
+    default_partitions : `int`, optional
+        The default number of Kafka partitions for each topic.
     queue_len : `int`, optional
         Length of the read queue.
 
@@ -93,16 +106,21 @@ class TopicNamesSet:
     """
 
     def __init__(
-        self, *, component, topic_names_list, queue_len=salobj.topics.DEFAULT_QUEUE_LEN
+        self,
+        *,
+        component,
+        topic_names_list,
+        default_partitions=1,
+        queue_len=salobj.topics.DEFAULT_QUEUE_LEN,
     ):
         self.component = component
+        self.default_partitions = default_partitions
         self.topic_names_list = list(topic_names_list)
         self.queue_len = int(queue_len)
-        topic_categories = ("commands", "events", "telemetry")
 
         # Construct a dict of topic category: all topic names, by reading the
         # IDL file. Topic names *omit* the command_ and logevent_ prefix.
-        all_names_dict = {category: set() for category in topic_categories}
+        all_names_dict = {category: set() for category in TOPIC_CATEGORIES}
         idl_metadata = salobj.parse_idl(
             component, idl.get_idl_dir() / f"sal_revCoded_{component}.idl"
         )
@@ -119,7 +137,7 @@ class TopicNamesSet:
         # Examine the topic_names_list, recording which topics have been seen
         # and looking for invalid and duplicate names.
         ackcmd_added = False
-        seen_names_dict = {category: set() for category in topic_categories}
+        seen_names_dict = {category: set() for category in TOPIC_CATEGORIES}
         for i, topic_names in enumerate(self.topic_names_list):
             if topic_names.add_ackcmd:
                 if ackcmd_added:
@@ -128,7 +146,7 @@ class TopicNamesSet:
                     )
                 ackcmd_added = True
 
-            for category in topic_categories:
+            for category in TOPIC_CATEGORIES:
                 all_names = all_names_dict[category]
                 seen_names = seen_names_dict[category]
                 specified_names = set(getattr(topic_names, category, []))
@@ -154,17 +172,21 @@ class TopicNamesSet:
         # Append a TopicNames to handle remaining topics, if needed.
         remaining_names = {
             category: all_names_dict[category] - seen_names_dict[category]
-            for category in topic_categories
+            for category in TOPIC_CATEGORIES
         }
         if not ackcmd_added or any(
-            len(remaining_names[category]) > 0 for category in topic_categories
+            len(remaining_names[category]) > 0 for category in TOPIC_CATEGORIES
         ):
             self.topic_names_list.append(
-                TopicNames(add_ackcmd=not ackcmd_added, **remaining_names)
+                TopicNames(
+                    partitions=self.default_partitions,
+                    add_ackcmd=not ackcmd_added,
+                    **remaining_names,
+                )
             )
 
     @classmethod
-    def from_file(cls, filename):
+    def from_file(cls, filename, default_partitions=1):
         """Create from a file path.
 
         Parameters
@@ -172,6 +194,8 @@ class TopicNamesSet:
         filename : `str`
             Path of yaml file to load.
             The schema must match that from `schema`
+        default_partitions : `int`, optional
+            The default number of Kafka partitions for each topic.
 
         Returns
         -------
@@ -189,10 +213,14 @@ class TopicNamesSet:
         topic_names_list = []
         component = components_info["component"]
         for topic_set in components_info.get("topic_sets", []):
+            topic_set.setdefault("partitions", default_partitions)
             topic_names_list.append(TopicNames(**topic_set))
         queue_len = components_info.get("queue_len", salobj.topics.DEFAULT_QUEUE_LEN)
         return cls(
-            component=component, topic_names_list=topic_names_list, queue_len=queue_len
+            component=component,
+            topic_names_list=topic_names_list,
+            default_partitions=default_partitions,
+            queue_len=queue_len,
         )
 
     @staticmethod
@@ -249,7 +277,15 @@ class TopicNamesSet:
                   type: array
                   items:
                     type: string
+                partitions:
+                    description: >-
+                        The number of Kafka partitions.
+                        If omitted, use the value specified by the
+                        --partitions command-line argument.
+                    type: integer
+                    minimum: 1
         required:
           - component
+          - topic_sets
         """
         )